+++ ydb/library/actors/subsystems/ya.make
@@ -14,18 +14,26 @@ IF (ALLOCATOR == "B" OR ALLOCATOR == "BS" OR ALLOCATOR == "C")
 ENDIF()
 
 SRCS(
-    stats.cpp
+    cgroup/cgroup_oom.cpp
+    cgroup/cgroup_oom_trend.cpp
+    cgroup/cgroup_v1.cpp
+    cgroup/cgroup_v2.cpp
 )
 
 PEERDIR(
-    ydb/library/actors/util
+    ydb/library/actors/core
     ydb/library/actors/protos
+    ydb/library/actors/util
 )
 
 IF (SANITIZER_TYPE == "thread")
     SUPPRESSIONS(
-        ../tsan.supp
+        ../core/tsan.supp
     )
 ENDIF()
 
 END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
