+++ ydb/library/actors/ya.make
@@ -9,6 +9,7 @@ RECURSE(
     log_backend
     memory_log
     prof
+    subsystems
     struct_log
     testlib
     util
